Wire in writeSource and remove prependFile from OutputMessages
