+++ unison-cli/src/Unison/Codebase/Editor/HandleInput.hs
@@ -68,6 +68,7 @@ import Unison.Codebase.Editor.HandleInput.BranchRename (handleBranchRename)
 import Unison.Codebase.Editor.HandleInput.Branches (handleBranches)
 import Unison.Codebase.Editor.HandleInput.DeleteBranch (handleDeleteBranch)
 import Unison.Codebase.Editor.HandleInput.DeleteProject (handleDeleteProject)
+import Unison.Codebase.Editor.HandleInput.FindAndReplace (handleStructuredFindI, handleStructuredFindReplaceI)
 import Unison.Codebase.Editor.HandleInput.Load (EvalMode (Sandboxed), evalUnisonFile, handleLoad, loadUnisonFile)
 import Unison.Codebase.Editor.HandleInput.MetadataUtils (addDefaultMetadata, manageLinks)
 import Unison.Codebase.Editor.HandleInput.MoveAll (handleMoveAll)
@@ -84,6 +85,7 @@ import Unison.Codebase.Editor.HandleInput.Projects (handleProjects)
 import Unison.Codebase.Editor.HandleInput.Pull (doPullRemoteBranch, mergeBranchAndPropagateDefaultPatch, propagatePatch)
 import Unison.Codebase.Editor.HandleInput.Push (handleGist, handlePushRemoteBranch)
 import Unison.Codebase.Editor.HandleInput.ReleaseDraft (handleReleaseDraft)
+import Unison.Codebase.Editor.HandleInput.Run (handleRun)
 import Unison.Codebase.Editor.HandleInput.RuntimeUtils qualified as RuntimeUtils
 import Unison.Codebase.Editor.HandleInput.TermResolution (resolveCon, resolveMainRef, resolveTermRef)
 import Unison.Codebase.Editor.HandleInput.Tests qualified as Tests
@@ -174,7 +176,7 @@ import Unison.Share.Codeserver qualified as Codeserver
 import Unison.ShortHash qualified as SH
 import Unison.Sqlite qualified as Sqlite
 import Unison.Symbol (Symbol)
-import Unison.Syntax.HashQualified qualified as HQ (fromString, toString, toText, toVar, unsafeFromString)
+import Unison.Syntax.HashQualified qualified as HQ (fromString, toString, toText, unsafeFromString)
 import Unison.Syntax.Lexer qualified as L
 import Unison.Syntax.Name qualified as Name (toString, toText, toVar, unsafeFromString, unsafeFromVar)
 import Unison.Syntax.Parser qualified as Parser
@@ -188,7 +190,6 @@ import Unison.Typechecker qualified as Typechecker
 import Unison.UnisonFile (TypecheckedUnisonFile)
 import Unison.UnisonFile qualified as UF
 import Unison.UnisonFile.Names qualified as UF
-import Unison.Util.Alphabetical qualified as Alphabetical
 import Unison.Util.Find qualified as Find
 import Unison.Util.List (nubOrdOn, uniqueBy)
 import Unison.Util.Monoid qualified as Monoid
@@ -203,8 +204,8 @@ import Unison.Util.TransitiveClosure (transitiveClosure)
 import Unison.Var (Var)
 import Unison.Var qualified as Var
 import Unison.WatchKind qualified as WK
+import UnliftIO.Directory qualified as Directory
 import Witch (unsafeFrom)
-import Unison.Codebase.Editor.HandleInput.Run (handleRun)
 
 ------------------------------------------------------------------------------------------------------------------------
 -- Main loop
@@ -1502,90 +1503,6 @@ inputDescription input =
       -- just trying to recover the syntax the user wrote
       These path _branch -> pure (Path.toText' path)
 
-handleStructuredFindI :: HQ.HashQualified Name -> Cli ()
-handleStructuredFindI rule = do
-  Cli.Env {codebase} <- ask
-  (ppe, names, rules0) <- lookupRewrite InvalidStructuredFind (\_ tm -> tm) rule
-  let rules = snd <$> rules0
-  let fqppe = PPED.unsuffixifiedPPE ppe
-  results :: [(HQ.HashQualified Name, Referent)] <- pure $ do
-    r <- Set.toList (Relation.ran $ Names.terms (NamesWithHistory.currentNames names))
-    Just hq <- [PPE.terms fqppe r]
-    fullName <- [HQ'.toName hq]
-    guard (not (Name.beginsWithSegment fullName Name.libSegment))
-    Referent.Ref _ <- pure r
-    Just shortName <- [PPE.terms (PPED.suffixifiedPPE ppe) r]
-    pure (HQ'.toHQ shortName, r)
-  let ok t@(_, Referent.Ref (Reference.DerivedId r)) = do
-        oe <- Cli.runTransaction (Codebase.getTerm codebase r)
-        pure $ (t, maybe False (\e -> any ($ e) rules) oe)
-      ok t = pure (t, False)
-  results0 <- traverse ok results
-  let results = Alphabetical.sortAlphabeticallyOn fst [(hq, r) | ((hq, r), True) <- results0]
-  let toNumArgs = Text.unpack . Reference.toText . Referent.toReference . view _2
-  #numberedArgs .= map toNumArgs results
-  Cli.respond (ListStructuredFind (fst <$> results))
-
-lookupRewrite :: (HQ.HashQualified Name -> Output) -> ([Symbol] -> Term Symbol Ann -> Term Symbol Ann) -> HQ.HashQualified Name -> Cli (PPED.PrettyPrintEnvDecl, NamesWithHistory, [(Term Symbol Ann -> Maybe (Term Symbol Ann), Term Symbol Ann -> Bool)])
-lookupRewrite onErr prepare rule = do
-  Cli.Env {codebase} <- ask
-  currentBranch <- Cli.getCurrentBranch0
-  hqLength <- Cli.runTransaction Codebase.hashLength
-  fileNames <- Cli.getNamesFromLatestParsedFile
-  let currentNames = NamesWithHistory.fromCurrentNames (fileNames <> Branch.toNames currentBranch)
-  let ppe = PPED.fromNamesDecl hqLength currentNames
-  ot <- Cli.getTermFromLatestParsedFile rule
-  ot <- case ot of
-    Just _ -> pure ot
-    Nothing -> do
-      case NamesWithHistory.lookupHQTerm NamesWithHistory.IncludeSuffixes rule currentNames of
-        s
-          | Set.size s == 1,
-            Referent.Ref (Reference.DerivedId r) <- Set.findMin s ->
-              Cli.runTransaction (Codebase.getTerm codebase r)
-        s -> Cli.returnEarly (TermAmbiguous (PPE.suffixifiedPPE ppe) rule s)
-  tm <- maybe (Cli.returnEarly (TermAmbiguous (PPE.suffixifiedPPE ppe) rule mempty)) pure ot
-  let extract vs tm = case tm of
-        Term.Ann' tm _typ -> extract vs tm
-        (DD.RewriteTerm' lhs rhs) ->
-          pure
-            ( ABT.rewriteExpression lhs rhs,
-              ABT.containsExpression lhs
-            )
-        (DD.RewriteCase' lhs rhs) ->
-          pure
-            ( Term.rewriteCasesLHS lhs rhs,
-              fromMaybe False . Term.containsCaseTerm lhs
-            )
-        (DD.RewriteSignature' _vs lhs rhs) ->
-          pure (Term.rewriteSignatures lhs rhs, Term.containsSignature lhs)
-        _ -> Cli.returnEarly (onErr rule)
-      extractOuter vs0 tm = case tm of
-        Term.Ann' tm _typ -> extractOuter vs0 tm
-        Term.LamsNamed' vs tm -> extractOuter (vs0 ++ vs) tm
-        tm -> case prepare vs0 tm of
-          DD.Rewrites' rules -> traverse (extract vs0) rules
-          _ -> Cli.returnEarly (onErr rule)
-  rules <- extractOuter [] tm
-  pure (ppe, currentNames, rules)
-
-handleStructuredFindReplaceI :: HQ.HashQualified Name -> Cli ()
-handleStructuredFindReplaceI rule = do
-  uf0 <- Cli.expectLatestParsedFile
-  let (prepare, uf, finish) = UF.prepareRewrite uf0
-  (ppe, _ns, rules) <- lookupRewrite InvalidStructuredFindReplace prepare rule
-  (dest, _) <- Cli.expectLatestFile
-  #latestFile ?= (dest, True)
-  let go n tm [] = if n == (0 :: Int) then Nothing else Just tm
-      go n tm ((r, _) : rules) = case r tm of
-        Nothing -> go n tm rules
-        Just tm -> go (n + 1) tm rules
-      (vs, uf0') = UF.rewrite (Set.singleton (HQ.toVar rule)) (\tm -> go 0 tm rules) uf
-      uf' = (vs, finish uf0')
-  #latestTypecheckedFile .= Just (Left . snd $ uf')
-  let msg = "| Rewrote using: "
-  Cli.respond $ OutputRewrittenFile ppe dest (msg <> HQ.toString rule) uf'
-
 handleFindI ::
   Bool ->
   FindScope ->
@@ -1909,16 +1826,12 @@ resolveHQToLabeledDependencies = \case
 
 doDisplay :: OutputLocation -> NamesWithHistory -> Term Symbol () -> Cli ()
 doDisplay outputLoc names tm = do
-  Cli.Env {codebase} <- ask
+  Cli.Env {codebase, writeSource} <- ask
   loopState <- State.get
 
   ppe <- prettyPrintEnvDecl names
   (tms, typs) <- maybe mempty UF.indexByReference <$> Cli.getLatestTypecheckedFile
-  let loc = case outputLoc of
-        ConsoleLocation -> Nothing
-        FileLocation path -> Just path
-        LatestFileLocation -> fmap fst (loopState ^. #latestFile) <|> Just "scratch.u"
-      useCache = True
+  let useCache = True
       evalTerm tm =
         fmap ErrorUtil.hush . fmap (fmap Term.unannotate) $
           RuntimeUtils.evalUnisonTermE True (PPE.suffixifiedPPE ppe) useCache (Term.amap (const External) tm)
@@ -1933,15 +1846,14 @@ doDisplay outputLoc names tm = do
       loadTypeOfTerm' (Referent.Ref (Reference.DerivedId r))
         | Just (_, _, ty) <- Map.lookup r tms = pure $ Just (void ty)
       loadTypeOfTerm' r = fmap (fmap void) . Cli.runTransaction . Codebase.getTypeOfReferent codebase $ r
-  rendered <-
-    DisplayValues.displayTerm
-      ppe
-      loadTerm
-      loadTypeOfTerm'
-      evalTerm
-      loadDecl
-      tm
-  Cli.respond $ DisplayRendered loc rendered
+  rendered <- DisplayValues.displayTerm ppe loadTerm loadTypeOfTerm' evalTerm loadDecl tm
+  mayFP <- case outputLoc of
+    ConsoleLocation -> pure Nothing
+    FileLocation path -> Just <$> Directory.canonicalizePath path
+    LatestFileLocation -> traverse Directory.canonicalizePath $ fmap fst (loopState ^. #latestFile) <|> Just "scratch.u"
+  whenJust mayFP \fp -> do
+    liftIO $ writeSource (Text.pack fp) (Cli.PrependSource False $ Text.pack . P.toPlain 80 $ rendered)
+  Cli.respond $ DisplayRendered mayFP rendered
 
 getLinks ::
   SrcLoc ->
