+++ unison-cli/src/Unison/CommandLine/OutputMessages.hs
@@ -9,7 +9,7 @@ import Control.Exception (catch, finally, mask, throwIO)
 import Control.Lens hiding (at)
 import Control.Monad.State
 import Control.Monad.State.Strict qualified as State
-import Control.Monad.Writer (Writer, mapWriter, runWriter, tell)
+import Control.Monad.Writer (Writer, runWriter, tell)
 import Data.ByteString.Lazy qualified as LazyByteString
 import Data.Foldable qualified as Foldable
 import Data.List (stripPrefix)
@@ -89,7 +89,6 @@ import Unison.ConstructorReference (GConstructorReference (..))
 import Unison.ConstructorType qualified as CT
 import Unison.Core.Project (ProjectBranchName (UnsafeProjectBranchName))
 import Unison.DataDeclaration qualified as DD
-import Unison.Debug qualified as Debug
 import Unison.Hash qualified as Hash
 import Unison.Hash32 (Hash32)
 import Unison.HashQualified qualified as HQ
@@ -105,7 +104,6 @@ import Unison.NamesWithHistory qualified as Names
 import Unison.Parser.Ann (Ann, startingLine)
 import Unison.Prelude
 import Unison.PrettyPrintEnv qualified as PPE
-import Unison.PrettyPrintEnv.Names qualified as PPE
 import Unison.PrettyPrintEnv.Util qualified as PPE
 import Unison.PrettyPrintEnvDecl qualified as PPED
 import Unison.PrettyTerminal
@@ -134,7 +132,6 @@ import Unison.Share.Sync.Types (CodeserverTransportError (..))
 import Unison.ShortHash qualified as ShortHash
 import Unison.Symbol (Symbol)
 import Unison.Sync.Types qualified as Share
-import Unison.Syntax.DeclPrinter (AccessorName)
 import Unison.Syntax.DeclPrinter qualified as DeclPrinter
 import Unison.Syntax.HashQualified qualified as HQ (toString, toText, unsafeFromVar)
 import Unison.Syntax.Name qualified as Name (toString, toText)
@@ -156,7 +153,6 @@ import Unison.Term (Term)
 import Unison.Term qualified as Term
 import Unison.Type (Type)
 import Unison.UnisonFile qualified as UF
-import Unison.UnisonFile.Names qualified as UF
 import Unison.Util.List qualified as List
 import Unison.Util.Monoid (intercalateMap)
 import Unison.Util.Monoid qualified as Monoid
@@ -769,7 +765,7 @@ notifyUser dir = \case
         ]
   DisplayDefinitions output -> displayDefinitions output
   DisplayDefinitionsString isTranscript definitions -> displayDefinitionsString isTranscript definitions
-  OutputRewrittenFile ppe dest msg uf -> displayOutputRewrittenFile ppe dest msg uf
+  OutputRewrittenFile dest vs -> displayOutputRewrittenFile dest vs
   DisplayRendered outputLoc pp ->
     displayRendered outputLoc pp
   TestResults stats ppe _showSuccess _showFailures oks fails -> case stats of
@@ -2513,13 +2509,11 @@ formatMissingStuff terms types =
              <> P.column2 [(P.syntaxToColor $ prettyHashQualified name, fromString (show ref)) | (name, ref) <- types]
        )
 
-displayOutputRewrittenFile :: (Ord a, Var v) => PPED.PrettyPrintEnvDecl -> FilePath -> String -> ([v], UF.UnisonFile v a) -> IO Pretty
-displayOutputRewrittenFile _ppe _fp _ ([], _uf) =
+displayOutputRewrittenFile :: (Var v) => FilePath -> [v] -> IO Pretty
+displayOutputRewrittenFile _fp [] =
   pure "😶️ I couldn't find any matches in the file."
-displayOutputRewrittenFile ppe fp msg (vs, uf) = do
+displayOutputRewrittenFile fp vs = do
   let modifiedDefs = P.sep " " (P.blue . prettyVar <$> vs)
-  let header = "-- " <> P.string msg <> "\n" <> "-- | Modified definition(s): " <> modifiedDefs
-  fp <- prependToFile (header <> "\n\n" <> prettyUnisonFile ppe uf <> foldLine) fp
   pure $
     P.callout "☝️" . P.lines $
       [ P.wrap $ "I found and replaced matches in these definitions: " <> modifiedDefs,
@@ -2529,49 +2523,6 @@ displayOutputRewrittenFile ppe fp msg (vs, uf) = do
 
 foldLine :: (IsString s) => P.Pretty s
 foldLine = "\n\n---- Anything below this line is ignored by Unison.\n\n"
-
-prettyUnisonFile :: forall v a. (Var v, Ord a) => PPED.PrettyPrintEnvDecl -> UF.UnisonFile v a -> Pretty
-prettyUnisonFile ppe uf@(UF.UnisonFileId datas effects terms watches) =
-  P.sep "\n\n" (map snd . sortOn fst $ prettyEffects <> prettyDatas <> catMaybes prettyTerms <> prettyWatches)
-  where
-    prettyEffects = map prettyEffectDecl (Map.toList effects)
-    (prettyDatas, accessorNames) = runWriter $ traverse prettyDataDecl (Map.toList datas)
-    prettyTerms = map (prettyTerm accessorNames) terms
-    prettyWatches = Map.toList watches >>= \(wk, tms) -> map (prettyWatch . (wk,)) tms
-
-    prettyEffectDecl :: (v, (Reference.Id, DD.EffectDeclaration v a)) -> (a, Pretty)
-    prettyEffectDecl (n, (r, et)) =
-      (DD.annotation . DD.toDataDecl $ et, st $ DeclPrinter.prettyDecl ppe' (rd r) (hqv n) (Left et))
-    prettyDataDecl :: (v, (Reference.Id, DD.DataDeclaration v a)) -> Writer (Set AccessorName) (a, Pretty)
-    prettyDataDecl (n, (r, dt)) =
-      (DD.annotation dt,) . st <$> (mapWriter (second Set.fromList) $ DeclPrinter.prettyDeclW ppe' (rd r) (hqv n) (Right dt))
-    prettyTerm :: Set (AccessorName) -> (v, a, Term v a) -> Maybe (a, Pretty)
-    prettyTerm skip (n, a, tm) =
-      if traceMember isMember then Nothing else Just (a, pb hq tm)
-      where
-        traceMember =
-          if Debug.shouldDebug Debug.Update
-            then trace (show hq ++ " -> " ++ if isMember then "skip" else "print")
-            else id
-        isMember = Set.member hq skip
-        hq = hqv n
-    prettyWatch :: (String, (v, a, Term v a)) -> (a, Pretty)
-    prettyWatch (wk, (n, a, tm)) = (a, go wk n tm)
-      where
-        go wk v tm = case wk of
-          WK.RegularWatch
-            | Var.UnnamedWatch _ _ <- Var.typeOf v ->
-                "> " <> P.indentNAfterNewline 2 (TermPrinter.pretty sppe tm)
-          WK.RegularWatch -> "> " <> pb (hqv v) tm
-          WK.TestWatch -> "test> " <> st (TermPrinter.prettyBindingWithoutTypeSignature sppe (hqv v) tm)
-          w -> P.string w <> "> " <> pb (hqv v) tm
-    st = P.syntaxToColor
-    sppe = PPED.suffixifiedPPE ppe'
-    pb v tm = st $ TermPrinter.prettyBinding sppe v tm
-    ppe' = PPED.PrettyPrintEnvDecl dppe dppe `PPED.addFallback` ppe
-    dppe = PPE.fromNames 8 (Names.NamesWithHistory (UF.toNames uf) mempty)
-    rd = Reference.DerivedId
-    hqv v = HQ.unsafeFromVar v
 
 displayDefinitions' ::
   (Var v) =>
@@ -2620,31 +2571,15 @@ displayDefinitions' ppe0 types terms = P.syntaxToColor $ P.sep "\n\n" (prettyTyp
 
 displayRendered :: Maybe FilePath -> Pretty -> IO Pretty
 displayRendered outputLoc pp =
-  maybe (pure pp) scratchAndDisplay outputLoc
+  pure $ maybe pp scratchMessage outputLoc
   where
-    scratchAndDisplay path = do
-      path' <- prependToFile pp path
-      pure (message pp path')
-      where
-        message pp path =
-          P.callout "☝️" $
-            P.lines
-              [ P.wrap $ "I added this to the top of " <> fromString path,
-                "",
-                P.indentN 2 pp
-              ]
-
--- Prepends the rendered Pretty to the file, returning the canonicalized FilePath
-prependToFile :: Pretty -> FilePath -> IO FilePath
-prependToFile pp path0 = do
-  path <- canonicalizePath path0
-  existingContents <- do
-    exists <- doesFileExist path
-    if exists
-      then readUtf8 path
-      else pure ""
-  writeUtf8 path . Text.pack . P.toPlain 80 $ pp <> P.text existingContents
-  pure path
+    scratchMessage path =
+      P.callout "☝️" $
+        P.lines
+          [ P.wrap $ "I added this to the top of " <> fromString path,
+            "",
+            P.indentN 2 pp
+          ]
 
 displayDefinitions :: DisplayDefinitionsOutput -> IO Pretty
 displayDefinitions DisplayDefinitionsOutput {isTest, outputFile, prettyPrintEnv = ppe, terms, types} =
